+++ NEWS.md
@@ -8,7 +8,8 @@
 * `sg_get_nodes_p` and `sg_get_edges_p` added to retrieve nodes and edges from draw graph.
 * `sg_drop_edges_p` and `sg_drop_nodes_p` now function properly.
 * The filter family of functions sees the addition of the `name` argument as well as `sg_filter_undo_p` to undo fitlers (by name).
-* `sg_clear` added
+* `sg_clear` added to clear the graph.
+* `sg_change_*_p` family added to change nodes and edges attributes on the fly.
 
 # sigmajs 0.1.2